requirements.txt
Django==3.2.6
djangorestframework==3.12.2
django-filter==2.4.0
markdown==3.3.4
django-ckeditor==1.2.2
drf-writable-nested==0.6.3